Extracted settings into a new component
